fix: Add `sort(nulls_last=(..., )` compat
